backend/requirements.txt
# GraphOps Backend Dependencies

# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# NebulaGraph
nebula3-python>=3.8.0

# Qdrant
qdrant-client>=1.12.0

# Redis
redis>=5.2.0

# Schema validation
pyyaml>=6.0.2
jsonschema>=4.23.0

# Excel parsing
openpyxl>=3.1.0

# File upload
python-multipart>=0.0.9

# Utilities
uuid7>=0.1.0
python-dotenv>=1.0.1
httpx>=0.28.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Job queue (installed now, used in M1+)
rq>=2.0.0
